Seperated the users and user_types routers
